feat: TNL-10136 minimum lines change
keep course enrollment changes to a minimum number of lines
